# Faster timestamp parsing (~70-90% faster)