--- a/arrow-cast/Cargo.toml
+++ b/arrow-cast/Cargo.toml
@@ -48,5 +48,10 @@ num = { version = "0.4", default-features = false, features = ["std"] }
 lexical-core = { version = "^0.8", default-features = false, features = ["write-integers", "write-floats", "parse-integers", "parse-floats"] }
 
 [dev-dependencies]
+criterion = { version = "0.4", default-features = false }
 
 [build-dependencies]
+
+[[bench]]
+name = "parse_timestamp"
+harness = false