# Which issue does this PR close?

<!--
We generally require a GitHub issue to be filed for all bug fixes and enhancements and this helps us generate change logs for our releases. You can link an issue to this PR using the GitHub syntax. For example `Closes #123` indicates that this PR will close issue #123.
-->

Closes #.

# Rationale for this change
 
<!--
Why are you proposing this change? If this is already explained clearly in the issue then this section is not needed.
Explaining clearly why changes are proposed helps reviewers understand your changes and offer better suggestions for fixes.
-->

```
2020-09-08              time:   [41.412 ns 41.433 ns 41.453 ns]
                        change: [-94.894% -94.888% -94.881%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe

2020-09-08T13:42:29     time:   [45.773 ns 45.807 ns 45.840 ns]
                        change: [-91.959% -91.950% -91.941%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 100 measurements (2.00%)
  1 (1.00%) low mild
  1 (1.00%) high mild

2020-09-08T13:42:29.190 time:   [46.387 ns 46.458 ns 46.534 ns]
                        change: [-91.947% -91.936% -91.925%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  4 (4.00%) high mild

2020-09-08T13:42:29.190855
                        time:   [47.343 ns 47.367 ns 47.393 ns]
                        change: [-91.397% -91.389% -91.382%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) low mild

2020-09-08T13:42:29.190855999
                        time:   [47.425 ns 47.468 ns 47.514 ns]
                        change: [-91.861% -91.851% -91.842%] (p = 0.00 < 0.05)
                        Performance has improved.

2020-09-08T13:42:29+00:00
                        time:   [57.554 ns 57.587 ns 57.623 ns]
                        change: [-32.570% -32.405% -32.206%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  1 (1.00%) high mild
  16 (16.00%) high severe

2020-09-08T13:42:29.190+00:00
                        time:   [59.645 ns 59.678 ns 59.711 ns]
                        change: [-33.385% -33.207% -33.058%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  3 (3.00%) low mild
  2 (2.00%) high mild
  1 (1.00%) high severe

2020-09-08T13:42:29.190855+00:00
                        time:   [57.440 ns 57.476 ns 57.515 ns]
                        change: [-36.681% -36.590% -36.504%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild

2020-09-08T13:42:29.190855999-05:00
                        time:   [57.124 ns 57.151 ns 57.179 ns]
                        change: [-37.896% -37.674% -37.515%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe

2020-09-08T13:42:29.190855Z
                        time:   [28.680 ns 28.712 ns 28.749 ns]
                        change: [-67.723% -67.672% -67.616%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  1 (1.00%) low mild
  7 (7.00%) high mild
  5 (5.00%) high severe
```

# What changes are included in this PR?

<!--
There is no need to duplicate the description in the issue here but it is sometimes worth providing a summary of the individual changes in this PR.
-->

# Are there any user-facing changes?


<!--
If there are user-facing changes then we may require documentation to be updated before approving the PR.
-->

<!---
If there are any breaking changes to public APIs, please add the `breaking change` label.
-->